pyproject.toml
[tool.poetry]
name = "nation-pycord"
version = "0.1.0"
description = "Pro Clubs Nation Discord Bot"
authors = ["compil3 <luca.tenuta@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
py-cord = {git = "https://github.com/naoTimesdev/pycord.git"}
python-dotenv = "^0.19.1"
pymongo = "^3.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
